requirements.txt
matplotlib==3.10.3
networkx==3.1
numpy==2.3.1
openml==0.15.1
pandas==2.3.0
Pillow==11.2.1
scikit_learn==1.7.0
scipy==1.15.3
tab_transformer_pytorch==0.4.2
torch==2.7.1
torchvision==0.22.1
vit_pytorch==1.10.1
